Eslint config: refactor internal structure
The purpose of this change is to modify internal structure of the
config. Specifically, it moves some files under `src/`, separates
presets into individual files, dedupes necessary plugins for each preset
and fixes some minor issues which were uncovered during this
refactoring.

There are still some parts that should be fixed (for example do not
include prettier in every preset), however, I wanted to keep the
snapshot changes on minimum so it's clear what changed.

I will follow with additional cleanup commits later.
