+++ src/eslint-config-adeira/__tests__/__fixtures__/README.md
