--- src/eslint-config-adeira/__tests__/eslintRules.js
@@ -1,16 +1,0 @@
-// @flow
-
-import { rules as builtInRules } from 'eslint/conf/eslint-all';
-
-const eslintRules: Set<string> = new Set(Object.keys(builtInRules));
-
-Object.keys(require('../package.json').dependencies)
-  .filter((dep) => dep.startsWith('eslint-plugin'))
-  .map((dep) => dep.replace('eslint-plugin-', ''))
-  .forEach((plugin) => {
-    Object.keys(require(`eslint-plugin-${plugin}`).rules).forEach((rule) => {
-      eslintRules.add(`${plugin}/${rule}`);
-    });
-  });
-
-export default eslintRules;
