+++ src/eslint-config-adeira/jest.js
@@ -1,20 +1,17 @@
 // @flow
 
-const { jest } = require('./presetsConfig');
-const changeNextVersionErrorLevel = require('./changeNextVersionErrorLevel');
-const getCommonConfig = require('./getCommonConfig');
+const changeNextVersionErrorLevel = require('./src/changeNextVersionErrorLevel');
+const getCommonConfig = require('./src/getCommonConfig');
+const jestPreset = require('./src/presets/jest');
+const { WARN } = require('./src/constants');
 
 /*::
 
-import type { EslintConfig } from './EslintConfig.flow';
+import type { EslintConfig } from './src/EslintConfig.flow';
 
 */
 
-const WARN = 1;
-
-const preset = ({
-  ...getCommonConfig(changeNextVersionErrorLevel(jest.rules, WARN)),
-  plugins: jest.plugins,
-} /*: EslintConfig */);
-
-module.exports = preset;
+module.exports = (getCommonConfig(
+  changeNextVersionErrorLevel(jestPreset.rules, WARN),
+  jestPreset.plugins,
+) /*: EslintConfig */);
