+++ src/eslint-config-adeira/src/EslintConfig.flow.js
@@ -3,20 +3,19 @@
 // Level "3" is our NEXT_VERSION_ERROR speciality:
 type EslintConfigErrorLevel = 0 | 1 | 2 | 'off' | 'warn' | 'error' | 3;
 
-type EslintRuleOption = string | { +[name: string]: mixed, ... };
+type EslintRuleOption = string | { +[name: string]: mixed };
 
 type EslintConfigValues =
   | EslintConfigErrorLevel
   | [EslintConfigErrorLevel, EslintRuleOption]
   | [EslintConfigErrorLevel, EslintRuleOption, EslintRuleOption];
 
-type EslintOverrides = $ReadOnlyArray<{ ... }>;
+export type EslintConfigRules = { +[string]: EslintConfigValues };
 
-export type EslintConfigRules = { +[string]: EslintConfigValues, ... };
 export type EslintConfig = {
-  +rules: EslintConfigValues,
-  +overrides: EslintOverrides,
+  +rules: EslintConfigRules,
   +plugins: $ReadOnlyArray<string>,
+  +overrides?: $ReadOnlyArray<{ ... }>,
   +settings?: { +[string]: $FlowFixMe },
-  +globals?: { +[string]: string },
+  +globals?: { +[string]: 'readonly' | 'writable' | 'off' },
 };
