+++ src/eslint-config-adeira/src/changeNextVersionErrorLevel.js
