+++ src/eslint-config-adeira/src/constants.js
@@ -1,0 +1,8 @@
+// @flow strict
+
+module.exports = {
+  OFF: 0,
+  WARN: 1,
+  ERROR: 2,
+  NEXT_VERSION_ERROR: 3, // special value changed later depending on the required strictness
+};
