+++ src/eslint-config-adeira/src/deprecatedRules.js
