+++ src/eslint-config-adeira/src/presets/base.js
@@ -1,22 +1,24 @@
-// @flow strict
+// @flow
+
+const { ERROR, NEXT_VERSION_ERROR, OFF, WARN } = require('../constants');
 
 /*::
 
-import type { EslintConfigRules } from './EslintConfig.flow';
+import type { EslintConfig } from '../EslintConfig.flow';
 
 */
 
-const OFF = 0;
-const WARN = 1;
-const ERROR = 2;
-const NEXT_VERSION_ERROR = 3; // special value changed later depending on the required strictness
-
-// Please note: there are some TODOs which means we will make it more strict
-// in the future versions. This change requires major version bump. It's not
-// safe to immediately switch from OFF to ERROR. Some time for migration is
-// needed (WARN serves this purpose).
-const groupedRules = ({
-  base: {
+module.exports = ({
+  plugins: [
+    'eslint-plugin-import',
+    'eslint-plugin-monorepo',
+    'eslint-plugin-node',
+    'eslint-plugin-eslint-comments',
+    'eslint-plugin-promise',
+    'eslint-plugin-adeira',
+    'eslint-plugin-sx',
+  ],
+  rules: {
     // Possible Errors (http://eslint.org/docs/rules/#possible-errors)
     'for-direction': ERROR,
     'getter-return': ERROR,
@@ -440,323 +442,4 @@ const groupedRules = ({
     'sx/use-logical-properties': NEXT_VERSION_ERROR,
     'sx/valid-usage': ERROR,
   },
-  jest: {
-    // Jest (https://github.com/jest-community/eslint-plugin-jest)
-    'jest/consistent-test-it': OFF,
-    'jest/expect-expect': OFF,
-    'jest/lowercase-name': OFF,
-    'jest/max-nested-describe': [NEXT_VERSION_ERROR, { max: 5 }],
-    'jest/no-alias-methods': OFF,
-    'jest/no-commented-out-tests': OFF,
-    'jest/no-conditional-expect': OFF,
-    'jest/no-deprecated-functions': ERROR,
-    'jest/no-disabled-tests': ERROR,
-    'jest/no-done-callback': OFF,
-    'jest/no-duplicate-hooks': ERROR,
-    'jest/no-expect-resolves': OFF,
-    'jest/no-export': ERROR,
-    'jest/no-focused-tests': ERROR,
-    'jest/no-hooks': OFF, // TODO: WARN (?)
-    'jest/no-identical-title': ERROR,
-    'jest/no-if': OFF,
-    'jest/no-interpolation-in-snapshots': ERROR,
-    'jest/no-jasmine-globals': ERROR,
-    'jest/no-jest-import': ERROR,
-    'jest/no-large-snapshots': OFF,
-    'jest/no-mocks-import': ERROR,
-    'jest/no-restricted-matchers': OFF,
-    'jest/no-standalone-expect': ERROR,
-    'jest/no-test-callback': OFF,
-    'jest/no-test-prefixes': OFF,
-    'jest/no-test-return-statement': ERROR,
-    'jest/no-truthy-falsy': OFF,
-    'jest/prefer-called-with': ERROR,
-    'jest/prefer-expect-assertions': OFF,
-    'jest/prefer-hooks-on-top': ERROR,
-    'jest/prefer-inline-snapshots': OFF,
-    'jest/prefer-spy-on': OFF,
-    'jest/prefer-strict-equal': OFF, // TODO: WARN
-    'jest/prefer-to-be-null': ERROR,
-    'jest/prefer-to-be-undefined': ERROR,
-    'jest/prefer-to-contain': ERROR,
-    'jest/prefer-to-have-length': ERROR,
-    'jest/prefer-todo': ERROR,
-    'jest/require-to-throw-message': OFF,
-    'jest/require-top-level-describe': OFF,
-    'jest/unbound-method': OFF,
-    'jest/valid-describe': ERROR,
-    'jest/valid-expect': ERROR,
-    'jest/valid-expect-in-promise': ERROR,
-    'jest/valid-title': ERROR,
-  },
-  flowtype: {
-    // flowtype (https://github.com/gajus/eslint-plugin-flowtype)
-    'flowtype/array-style-complex-type': OFF,
-    'flowtype/array-style-simple-type': OFF,
-    'flowtype/arrow-parens': OFF,
-    'flowtype/define-flow-type': WARN,
-    'flowtype/enforce-line-break': OFF,
-    'flowtype/interface-id-match': OFF,
-    'flowtype/newline-after-flow-annotation': [ERROR, 'always'],
-    'flowtype/no-dupe-keys': ERROR,
-    'flowtype/no-existential-type': ERROR, // https://github.com/facebook/flow/issues/6308
-    'flowtype/no-flow-fix-me-comments': OFF,
-    'flowtype/no-internal-flow-type': ERROR,
-    'flowtype/no-mixed': OFF,
-    'flowtype/no-mutable-array': OFF,
-    'flowtype/no-primitive-constructor-types': WARN,
-    'flowtype/no-types-missing-file-annotation': ERROR,
-    'flowtype/no-unused-expressions': [ERROR, { allowTaggedTemplates: true }],
-    'flowtype/no-weak-types': [
-      ERROR,
-      {
-        // codeshift: https://github.com/facebook/flow/issues/7291
-        any: false,
-        Object: true, // { +[key: string]: any, ... }
-        Function: true, // (...args: $ReadOnlyArray<any>) => any   -OR-   { [[call]]: any, ... }
-      },
-    ],
-    'flowtype/require-compound-type-alias': OFF,
-    'flowtype/require-exact-type': [ERROR, 'never'], // we are using `exact_by_default=true`
-    'flowtype/require-indexer-name': OFF,
-    'flowtype/require-inexact-type': OFF,
-    'flowtype/require-parameter-type': OFF,
-    'flowtype/require-readonly-react-props': [
-      ERROR,
-      {
-        useImplicitExactTypes: true,
-      },
-    ],
-    'flowtype/require-return-type': OFF,
-    'flowtype/require-types-at-top': OFF,
-    'flowtype/require-valid-file-annotation': [ERROR, 'always'],
-    'flowtype/require-variable-type': OFF,
-    'flowtype/sort-keys': OFF,
-    'flowtype/spread-exact-type': OFF, // TODO: WARN?
-    'flowtype/type-id-match': OFF,
-    'flowtype/type-import-style': OFF,
-    'flowtype/use-flow-type': WARN,
-    'flowtype/use-read-only-spread': ERROR,
-    'flowtype/valid-syntax': OFF,
-
-    // Flow FB: https://github.com/facebook/flow/tree/master/packages/eslint-plugin-fb-flow
-    'fb-flow/use-indexed-access-type': OFF, // TODO (revert https://github.com/adeira/universe/pull/2662)
-  },
-  react: {
-    // React (https://github.com/yannickcr/eslint-plugin-react)
-    'react/boolean-prop-naming': OFF,
-    'react/button-has-type': [ERROR, { button: true, submit: true, reset: false }],
-    'react/default-props-match-prop-types': OFF,
-    'react/destructuring-assignment': OFF,
-    'react/display-name': OFF,
-    'react/forbid-component-props': OFF,
-    'react/forbid-dom-props': [
-      ERROR,
-      {
-        forbid: [
-          'class', // https://reactjs.org/docs/dom-elements.html#classname
-          'for', // https://reactjs.org/docs/dom-elements.html#htmlfor
-        ],
-      },
-    ],
-    'react/forbid-elements': OFF,
-    'react/forbid-foreign-prop-types': OFF,
-    'react/forbid-prop-types': OFF,
-    'react/jsx-boolean-value': OFF,
-    'react/jsx-curly-brace-presence': OFF,
-    'react/jsx-filename-extension': OFF,
-    'react/jsx-fragments': OFF,
-    'react/jsx-handler-names': OFF,
-    'react/jsx-key': ERROR,
-    'react/jsx-max-depth': OFF,
-    'react/jsx-no-bind': [
-      ERROR,
-      {
-        ignoreDOMComponents: true,
-        ignoreRefs: true,
-        allowArrowFunctions: true,
-        allowFunctions: false,
-        allowBind: false,
-      },
-    ],
-    'react/jsx-no-comment-textnodes': ERROR,
-    'react/jsx-no-constructed-context-values': ERROR,
-    'react/jsx-no-duplicate-props': ERROR,
-    'react/jsx-no-literals': OFF,
-    'react/jsx-no-target-blank': ERROR,
-    'react/jsx-no-undef': ERROR,
-    'react/jsx-no-useless-fragment': ERROR,
-    'react/jsx-pascal-case': ERROR,
-    'react/jsx-props-no-spreading': OFF,
-    'react/jsx-sort-default-props': OFF,
-    'react/jsx-sort-props': OFF,
-    'react/jsx-uses-react': WARN,
-    'react/jsx-uses-vars': ERROR,
-    'react/no-access-state-in-setstate': ERROR,
-    'react/no-array-index-key': OFF,
-    'react/no-children-prop': OFF,
-    'react/no-danger': OFF,
-    'react/no-danger-with-children': ERROR,
-    'react/no-deprecated': ERROR,
-    'react/no-did-mount-set-state': OFF, // https://github.com/airbnb/javascript/issues/684#issuecomment-264094930
-    'react/no-did-update-set-state': ERROR,
-    'react/no-direct-mutation-state': ERROR,
-    'react/no-find-dom-node': ERROR,
-    'react/no-is-mounted': ERROR,
-    'react/no-multi-comp': [ERROR, { ignoreStateless: true }],
-    'react/no-redundant-should-component-update': ERROR,
-    'react/no-render-return-value': ERROR,
-    'react/no-set-state': OFF,
-    'react/no-string-refs': OFF,
-    'react/no-this-in-sfc': ERROR,
-    'react/no-typos': OFF,
-    'react/no-unescaped-entities': ERROR,
-    'react/no-unknown-property': ERROR,
-    'react/no-unsafe': OFF,
-    'react/no-unstable-nested-components': [ERROR, { allowAsProps: true }],
-    'react/no-unused-prop-types': OFF,
-    'react/no-unused-state': OFF, // Enable when they fix this issue: https://github.com/yannickcr/eslint-plugin-react/issues/1910
-    'react/no-will-update-set-state': OFF,
-    'react/prefer-es6-class': [ERROR, 'always'],
-    'react/prefer-read-only-props': OFF,
-    'react/prefer-stateless-function': OFF,
-    'react/prop-types': OFF, // we use Flow instead,
-    'react/react-in-jsx-scope': OFF,
-    'react/require-default-props': OFF,
-    'react/require-optimization': OFF,
-    'react/require-render-return': ERROR,
-    'react/self-closing-comp': ERROR,
-    'react/sort-comp': [
-      ERROR,
-      {
-        // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/sort-comp.md
-        order: [
-          'type-annotations',
-          'static-variables',
-          'static-methods',
-          'instance-variables',
-          'lifecycle',
-          'everything-else',
-          'rendering',
-        ],
-        groups: {
-          lifecycle: [
-            'displayName',
-            'propTypes',
-            'contextTypes',
-            'childContextTypes',
-            'mixins',
-            'statics',
-            'defaultProps',
-            'constructor',
-            'getDefaultProps',
-            'state',
-            'getInitialState',
-            'getChildContext',
-            'getDerivedStateFromProps', // this is added
-            'getDerivedStateFromError', // this is added
-            'componentWillMount',
-            'UNSAFE_componentWillMount',
-            'componentDidMount',
-            'componentWillReceiveProps',
-            'UNSAFE_componentWillReceiveProps',
-            'shouldComponentUpdate',
-            'componentWillUpdate',
-            'UNSAFE_componentWillUpdate',
-            'getSnapshotBeforeUpdate',
-            'componentDidUpdate',
-            'componentDidCatch',
-            'componentWillUnmount',
-          ],
-          rendering: ['/^render.+$/', 'render'],
-        },
-      },
-    ],
-    'react/sort-prop-types': OFF,
-    'react/state-in-constructor': OFF,
-    'react/static-property-placement': OFF,
-    'react/style-prop-object': ERROR,
-    'react/void-dom-elements-no-children': ERROR,
-    'react/function-component-definition': OFF,
-    'react/jsx-no-script-url': ERROR,
-    'react/no-adjacent-inline-elements': OFF,
-
-    // React Hooks (https://www.npmjs.com/package/eslint-plugin-react-hooks)
-    'react-hooks/rules-of-hooks': ERROR,
-    'react-hooks/exhaustive-deps': [
-      ERROR,
-      {
-        // https://recoiljs.org/docs/introduction/installation/#eslint
-        additionalHooks: 'useRecoilCallback', // should be regex!
-      },
-    ],
-
-    // React Native (https://github.com/intellicode/eslint-plugin-react-native)
-    'react-native/no-color-literals': OFF,
-    'react-native/no-inline-styles': OFF,
-    'react-native/no-raw-text': OFF, // TODO: takes normal React into account as well, not only RN
-    'react-native/no-single-element-style-arrays': ERROR,
-    'react-native/no-unused-styles': ERROR,
-    'react-native/sort-styles': OFF,
-    'react-native/split-platform-components': OFF,
-
-    // React Accessibility (https://github.com/evcohen/eslint-plugin-jsx-a11y)
-    'jsx-a11y/accessible-emoji': ERROR,
-    'jsx-a11y/alt-text': ERROR,
-    'jsx-a11y/anchor-has-content': ERROR,
-    'jsx-a11y/anchor-is-valid': ERROR,
-    'jsx-a11y/aria-activedescendant-has-tabindex': OFF,
-    'jsx-a11y/aria-props': ERROR,
-    'jsx-a11y/aria-proptypes': ERROR,
-    'jsx-a11y/aria-role': ERROR,
-    'jsx-a11y/aria-unsupported-elements': ERROR,
-    'jsx-a11y/autocomplete-valid': ERROR,
-    'jsx-a11y/click-events-have-key-events': OFF,
-    'jsx-a11y/control-has-associated-label': OFF,
-    'jsx-a11y/heading-has-content': ERROR,
-    'jsx-a11y/html-has-lang': ERROR,
-    'jsx-a11y/iframe-has-title': ERROR,
-    'jsx-a11y/img-redundant-alt': OFF,
-    'jsx-a11y/interactive-supports-focus': ERROR,
-    'jsx-a11y/label-has-associated-control': OFF,
-    'jsx-a11y/label-has-for': OFF,
-    'jsx-a11y/lang': ERROR,
-    'jsx-a11y/media-has-caption': OFF,
-    'jsx-a11y/mouse-events-have-key-events': OFF,
-    'jsx-a11y/no-access-key': ERROR,
-    'jsx-a11y/no-autofocus': [ERROR, { ignoreNonDOM: true }],
-    'jsx-a11y/no-distracting-elements': [ERROR, { elements: ['marquee', 'blink'] }],
-    'jsx-a11y/no-interactive-element-to-noninteractive-role': OFF,
-    'jsx-a11y/no-noninteractive-element-interactions': OFF,
-    'jsx-a11y/no-noninteractive-element-to-interactive-role': OFF,
-    'jsx-a11y/no-noninteractive-tabindex': ERROR,
-    'jsx-a11y/no-onchange': OFF,
-    'jsx-a11y/no-redundant-roles': ERROR,
-    'jsx-a11y/no-static-element-interactions': OFF,
-    'jsx-a11y/role-has-required-aria-props': ERROR,
-    'jsx-a11y/role-supports-aria-props': ERROR,
-    'jsx-a11y/scope': ERROR,
-    'jsx-a11y/tabindex-no-positive': ERROR,
-  },
-  relay: {
-    // Relay (https://github.com/relayjs/eslint-plugin-relay)
-    'relay/compat-uses-vars': OFF, // we do not use Relay Compat
-    'relay/function-required-argument': ERROR,
-    'relay/generated-flow-types': ERROR,
-    'relay/graphql-naming': OFF, // no longer needed, see: https://github.com/facebook/relay/commit/ff1c10bc6595f3715f29660b46f779e000be9c70
-    'relay/graphql-syntax': ERROR,
-    'relay/hook-required-argument': ERROR,
-    'relay/must-colocate-fragment-spreads': ERROR,
-    'relay/no-future-added-value': ERROR,
-    'relay/unused-fields': ERROR,
-  },
-} /*: {[string]:EslintConfigRules} */);
-
-exports.groupedRules = groupedRules;
-
-let ourRules /*: EslintConfigRules */ = {};
-Object.values(groupedRules).forEach((rules) => {
-  ourRules = { ...ourRules, ...rules };
-});
-
-exports.ourRules = ourRules;
+} /*: EslintConfig */);
